feat: Add weaviate client initializer
get_weaviate_client() needs WeaviateSettings that
set from .env file and it returns weaviate.WeaviateClient.

it raises WeaviateConnectionError when it failed to connect
weaviate and if client is not ready it raises WeaviateNotReadyError

create_vectorwave_schema() needs weaviate.WeaviateClient
and it generate vector db's columns

it raises SchemaCreationError when it got something wrong
while creating schema.

Related to #2

Signed-off-by: Jun Yeong Kim <junyeonggim5@gmail.com>